Add uuid field to SecurityGroup object
We're going to store requested security group objects in the
request spec. nova-network security groups are unique by name
and tenant. Neutron security groups, however, are not unique
by name within a tenant, they are unique by uuid. So to uniquely
store the requested security groups in the request spec we need
to store the neutron-specific id value (which is a uuid) in the
SecurityGroup.uuid field added here.

Since we don't store this in the nova database the uuid value
must be popped during a save() and ignored when creating an
object from a DB primitive.

Part of blueprint cells-scheduling-interaction

Change-Id: I3c3fb8fe9a53155caba49b353c0524f84dc458ff