placement: genericize on resource providers
A simple patch that makes a number of the methods of the scheduler
report client just take a resource provider UUID and a set of inventory
records. With upcoming patches, the compute node is not the only
provider of resources in the system and we want to keep various report
client methods for adding resource provider and inventory records as
generic as possible instead of being specific to a compute node.

Conflicts:
        nova/scheduler/client/report.py

NOTE(mriedem): The conflict is due to not having hacking change
4eb89c206e68a7172ebad897ad24769036c7bdd6 in stable/newton.

NOTE(mriedem): I've removed the blueprint reference from the commit
message on this change because while it was part of a series in Ocata,
it's a refactor patch which has no functional difference in how the
report client works and makes cherry-picking other changes on top of
this for actual bug fixes much simpler.

Change-Id: Ia8fe9d71893e8bf3f4f8cddd329613e269fef210
(cherry picked from commit f1cd928bf5daac3b344d0c6cd095aabec14cdbff)