Fix mock arg list order in test_driver.py
When we use nest patch decorators the mocks, they are passed in to the
decorated function in the same order they applied, this means should be
from the bottom up. Though We have never use these arguments in this
function and the order may not very important, it's a little misleading.
So I think we should fix it.

Change-Id: I459780a17998382b61f56377bfa17651a5f99f51