Avoid unnecessary db_calls in objects.Instance._from_db_object()
During construction of an instance object from db info, a call to
_from_db_object() is made. In many situations this results in one or
more unnecessary db calls due to the way instance extras are handled.

This occurs when the following two conditions apply; (1) the
'expected_attrs' parameter contains one of the affected instance_extras
fields (see below) and (2) the corresponding value in the provided
db_instance data either does not exist or else contains None.

The affected instance extras fields are:
 - numa_topology
 - pci_requests
 - device_metadata
 - vcpu_model

This change works around this by making use of the already defined
_NO_DATA_SENTINEL object to differentiate between a None value and no
value at all so that only the latter results in calling out to the db.

Change-Id: Ib09f3ba50ea2ec13d9d842923449b2f7faf32ff2
Closes-Bug: #1641413