rt: pass nodename to internal methods
We would like to remove the resource tracker's nodename attribute and
have the resource tracker treat Ironic baremetal resource nodes like any
other consumable resource in the system. This means having the Ironic
virt driver register Ironic baremetal nodes as resource providers in the
new placement API and creating an inventory record of a quantity of 1
with a resource class of the Ironic node.resource_class attribute. The
problem we have is that currently the nova-compute daemon keeps a
ResourceTracker object in memory for each Ironic compute node in the
entire deployment (or a segment of Ironic nodes if using Newton and the
Ironic hashring functionality). We would like to have a single resource
tracker object that manages a set of resource providers instead of
having a resource tracker for each Ironic baremetal node.

This patch performs some initial surgery on the resource tracker to pass
a nodename parameter to a number of methods.  The first one up is the
simple internal methods:

- _set_instance_host_and_nodename()
- _create_migration()
- _claim_existing_migraton()
- _move_claim()

Once the internal methods are cleaned, we will modify the public methods
called by the nova.compute.manager.Manager object to pass a nodename
parameter and remove the resource tracker's internal nodename attribute.
This will allow us to only have a single ResourceTracker object instance
(and thus a single scheduler reporting client object) on the
nova-compute daemon. The net result of this should be easier to reason
about resource tracking, more consistency in how resources are accounted
for, and a smaller memory-footprint for nova-compute when running Ironic
deployments.

Change-Id: I1b99fa48cc7902887a939d25097704eab6bb9c7a
blueprint: custom-resource-classes