rt: pass the nodename to public methods
We want to eventually stop treating the Ironic nodename as a "special"
kind of unicorn resource in the system. To get there, we need to
decouple the (host, nodename) state in the nova-compute resource
tracker. And in order to do that, we first need to stop creating a
resource tracker instance on the nova-compute for every Ironic bare
metal resource node.

In this patch we pass a nodename parameter to the public methods of the
resource tracker and use that parameter instead of the RT's nodename
attribute.

This allows us in the next patch to remove the .nodename attribute of
the resource tracker and use a single resource tracker instance in the
nova-compute manager that manages all compute node records. Once that is
done, we can move forward to tackling the Ironic custom resource class
and resource provider accounting.

Change-Id: I3bf31a4334ca50516af28930fcb452f50474878b
blueprint: custom-resource-classes