api-ref: body verification for force_complete server migration
This completes the body verification for the force_complete
action for server migrations.

This is only supported with microversion >= 2.22 and only for
in-progress live migrations, on unlocked servers that are in
active/migrating status.

There is also a note as not all compute drivers support this
action. See bug 1641753 for details.

Part of blueprint api-ref-in-rst-ocata

Change-Id: I9cd6cf35b4b5828f0f4acde168cda2beedd902f4