raise exception if create Virtuozzo container with swap disk
Virtuozzo containers based on OS-level virtualization.
OS containers don't use swap as additional disk.
So Virtuozzo doesn't support swap in terms of OpenStack.
We should fail "nova boot" operation if
user try to create Virtuozzo container which is based on flavor with swap disk.

In this patch we move the code which checks swap disk
to the beginning of the function.
We should check swap disk before creating ephemeral and root disk, because
in case of error we don't need to delete created disks.

Closes-Bug: 1696697

Change-Id: I1a17d9af3050fe70d49adfb7e65d6b292b6a52f5