Add a retry loop to ResourceClass creation
Since the integer IDs for custom resource classes are generated in code,
there is the possibility of a race condition when adding more than one
such class in a short period of time. This patch adds a loop to catch
the specific exception in this case, so that it can be retried.

Blueprint: custom-resource-classes

Change-Id: I15ad20333fb9724224bcb96427e0819ab5efd5bd