libvirt: Use the mirror element to detect job completion
The mirror element was introduced in Libvirt 0.10.0 [1] and provides a
ready attribute that helps to determine when a blockjob is complete.
This is now used in addition to checking the progress of a given
blockjob to ensure Nova waits until these jobs complete fully.

[1] https://libvirt.org/git/?p=libvirt.git;a=commit;h=ae6aa8c3965e9aaa245b8e669c6324d44312ac1b

Closes-bug: #1643017
Change-Id: I0c52917a5555a70c4973f37dea1aebf878dd73b4
(cherry picked from commit 9ed782641970dd371adc151e7d3913a402369dd3)