Add TODO for returning a 202 from the volume attach API
Attaching a volume is an RPC cast from the API node to the
compute node, so we should return a 202 instead of a 200
in the response. Making that change would break Tempest though
which expects a 200 response code, and we have several other
APIs with similar response code issues:

http://paste.openstack.org/show/590154/

So this just adds a TODO for the same bucket of issues which
we'll probably need to handle in a single microversion.

Change-Id: I0bee2ea8681cc001dd490607213bc987ec70301f