guestfs: Don't report exception if there's read access to kernel
Commit 92ae0f1 ("libvirt - Add log if libguestfs can't read host
kernel") reworks the logic of handling access to Kernel for libguestfs.
In doing that, it erroneously raises an exception when libguestfs is
_able_ to access the Kernel.

Fix it by reporting exception only when libguestfs does _not_ have
read access to the Kernel.

This was first tried by Kevin Zhao here
Ic6802650cb8f93e0d02c51e9014eb85a7e71f6fe, but is abandoned for some
reason.

This also adds a little more direction on what to do to fix this error.

Related-Bug: #1646002

Change-Id: Id6b4108e4e4af7c98b3e1bd9de3a09ad057e7b92