[scheduler][tests]: Fix incorrect aggr mock values
Currently the value return from 'utils.aggregate_metadata_get_by_host'
is a type of collections.defaultdict(set). That said all the values
are 'set' not 'String'. Unfortunately, most of test cases related to
aggregate filters are using it with in incorrect way. This patch fixes
it and adds an extra test case to verify it.

Co-Authored-By: Fei Long Wang <flwang@catalyst.net.nz>

Change-Id: I12f033a66ed31cd624bda47be944bab6c841dbf5