Make resource provider objects not remotable
The intent is for the resource provider objects to never be used
outside the placement API and for the placement HTTP API to be the
only remote interface to the information represented by the objects.
Therefore there is no need for the versioned objects that provide
the interface between the HTTP API and the database API to be
remotable.

This change removes the remotable decorators and updates tests
accordingly. The objects which have changed have had their versions
updated to provide a marker in their history to indicate that it
happened.

Change-Id: Ic2b431c8ba0af2dea4117687c7ef9477a6655bcf