Convert live migration uri back to string
If live_migration_inbound_addr is set in nova.conf, value of the option
will be saved under target_connect_addr in migrate_data. Unfortunately,
the type of value saved in target_connect_addr is unicode, therefore
when creating URI for live migration over dedicated interface we are
ending up with unicode type which is not acceptable by libvirt. To fix
the issue let's convert URI explictly back to string.

Change-Id: I651cc7a45fe06a2b6d81e007c4faf62947ee3203
Closes-Bug: #1638625