placement: report client handle InventoryInUse
In a Newton deployment, Ironic baremetal nodes are added as resource
providers to the placement service. Those resource providers will
typically have 3 inventory records, one each for VCPU, MEMORY_MB, and
DISK_GB. In an Ocata deployment, the Nova resource tracker is going to
switch to tracking Ironic baremetal nodes having a single inventory
record corresponding to the Ironic node's "resource_class" attribute (a
custom resource class).

When this happens, we will need to rework allocation records against
that baremetal node provider since we will not be able to delete the
inventory records that refer to the old resource classes like VCPU,
MEMORY_MB and DISK_GB until those old allocation records are removed.

This patch catches the 409 Conflict returned from the placement service
when an InventoryInUse exception occurs and re-raises this exception on
the client side, allowing us a way in future patches to differentiate
between a ConcurrentUpdateDetected and the issue described above.

Change-Id: Ie52e495945b7b3db5ae777804c50260e3324e0ea
blueprint: custom-resource-classes