Resource tracker doesn't free resources on confirm resize
If the audit hasn’t been triggered and confirm resize is executed, the
resources aren't dropped right away because the itype stored in the
self.tracked_migrations corresponds to the new flavor.  However, the
resource usage would get recalculated correctly by the next periodic
resource audit.

If the audit got executed, it corresponds to the old flavor, and resources
get dropped properly.

This patch removes part of the condition in drop_move_claim() which was
trying to validate that the instance type (itype) stored in the
tracked_migration was corresponding instance type being passed in the
method.  Looking at the code where drop_move_claim() is being called shows
that we can safely remove this check since it is known (based on the prefix)
which resources needs to be dropped.

Also, unit test changes from b816e3 were moved to a separate test.  It
introduced a hacky way to force the migration dest_compute and makes it hard
to keep _test_finish_revert_resize() generic and have the resources correctly
tracked.

Closes-Bug: #1647370
Change-Id: Ic02b56be5684d9206ece4c20f5194802dfe12e25