placement: minor refactor _allocate_for_instance()
The _allocate_for_instance() method of the scheduler report client was
designed to query the supplied Instance object and construct a payload
of allocation data to send to the placement service that only included
the VCPU, MEMORY_MB and DISK_GB resource classes.

In upcoming patches, we will be changing the way that Ironic baremetal
nodes have their inventory and allocations registered. Specifically, for
those instances that are assigned to an Ironic node, we will no longer
be "consuming" VCPU, MEMORY_MB and DISK_GB resources on an Ironic
baremetal node but rather, the instance will be consuming the node
itself in its entirety. The inventory of the baremetal node resource
provider will be a single record of an amount of 1 for the resource
class matching the Ironic node's "resource_class" attribute.

In preparation for this upcoming change, this patch splits the code in
the _allocate_for_instance() method that constructs the payload sent to
the placement service from the code that looks at the Instance object's
flavor representation's vcpus, memory_mb and local_gb amounts.

Change-Id: Id60f84253fdbadd3d2eb4bb66ea0d4b05cdfe752
blueprint: custom-resource-classes