support polling free notification testing
There are multiple instance action notifications where the only
way to test notification sample is to wait for the notfication
to arrive as there is no indication on the REST API that the
certain action has been finished.

Until now the only way to implement such test was to poll the
fake_notifier.VERSIONED_NOTIFICATIONS list to see if the notification
is arrived.

This patch adds a polling free solution based on threading.Event

Change-Id: I50f2a3325e3aae4298b18f0929d19c04cecdaa18