Support filtering resource providers by aggregate membership
This provides the backend for allowing an HTTP query of
'/resource_providers?member_of=in:$uuid_1,$uuid_2,$uuid_3'
which will return those resource providers which are members of any
of the listed aggregates.

This is done by extending the ResourceProvider.get_all_by_filters
methods to accept an additional 'member_of' filter with a value
of a list of aggregate uuids.

Change-Id: I1a69f7426b23f7b118bac7877b64700e688e8c6d