Fix crashing during guest config with pci_devices=None
The Instance.pci_devices field is nullable, but the get_instance_pci_devs()
function does not account for that. If it is passed an instance (from an
older deployment) with no pci_devices, it will crash trying to iterate
None as a list.

Change-Id: I3d535e01ac31db7804347c3938c0d88a28ba67f5
Closes-Bug: #1648887
(cherry picked from commit 15564eb355657fcf2594d5060cb36f25e28bc0c9)