Make simple_cell_setup fully idempotent
Previously this command would treat the presence of a cell0 mapping
as indication that cellsv2 is fully setup. This could be a false
indication for a variety of situations, including having set up a
cell0 mapping manually, as well as failing to sync the database and
thus not continuing on to host/instance mapping.

This patch makes us continue on, even if the mapping is present,
and also even if the sync of the cell0 database fails.

Change-Id: Ic05e1a998d7807d58fc35637742059d61f7669e9
Closes-Bug: #1649341