Let nova-manage cell_v2 commands use transport_url from CONF
The nova-manage cell_v2 simple_cell_setup command currently requires
a --transport-url to be passed on the command line. It should be
able to use the [DEFAULT]/transport_url setting from nova.conf if
present.

This makes --transport-url optional and uses it if specified. If not
specified, it looks at [DEFAULT]/transport_url, and finally raises
an error if neither were found.

Change-Id: Idfda66d161545a20f3102c083a7198720bf54e6b