Remove unused validation code from block_device
BDM code had some input validation code, however that is not used at
all because json-schema has validated the input before passing the
input to BDM code.
For example, source_type and boot_index are validated with

 'source_type': {
     'type': 'string',
     'enum': ['volume', 'image', 'snapshot', 'blank'],
 },
 'boot_index': {
     'type': ['integer', 'string', 'null'],
     'pattern': '^-?[0-9]+$',
 },

So Nova doesn't need to have the BDM code and this patch removes it.
This patch adds the corresponding test.

Change-Id: I278baead330c32d39f3d6c42077e3563205eed35