libvirt: Remove slowpath listing of instances
Commit 'e495d139' introduced a "fast path" method for listing all
domains. This requires libvirt 0.9.13 [1] but, as the minimum version of
libvirt at the time was 0.9.6 [2], the older "slow path" method was
retained.

Seeing as nova has long since moved on from 0.9.13, this "slow path" can
be dropped and the tests simplified to handle the single path remaining.
The (now unused) '_get_domain_by_id' function and its respective tests
are obliterated in the process.

This also fixes the 'listAllDomains' function of fakelibvirt.Connection,
which appears not to have been called previously and was broken.

[1] https://github.com/openstack/nova/blob/e495d139/nova/virt/libvirt/driver.py#L774
[2] https://github.com/openstack/nova/blob/e495d139/nova/virt/libvirt/driver.py#L289

Change-Id: I6a01529b4b98f5d29c90a7ba38da7e9337dae99b