Restrict CONF.quota.driver to DB and noop quota drivers
The quota driver config option was deprecated in Newton via
change 430638888c987a99e537e2ac956087a7310ecdc6. That was a
bit wrong in that the config option shouldn't have been
deprecated since we still use it, but we wanted to deprecate
the ability to load custom quota driver classes, which is done
in this change.

To be clear, this doesn't remove the option nor the ability
to configure the quota driver, it only removes the ability
to class-load out of tree quota drivers.

Change-Id: I021a2bcb923739409b393cbb2684ffdf20180f73