Revert "Use liberty-eol tag for liberty release notes"
This reverts commit 4a237c02e43dbefb61a1b1f34b3e268593b5061a

reno 1.9.0 has a bug which doesn't handle using an -eol tag
as the branch, because it transforms it blindly into something
like liberty-eol-eol which doesn't exist.

As we're coming up on the ocata-2 milestone and need to get
some release notes in for the beta tag, we need to just
revert this change to make things work with reno 1.9.0.

Change-Id: I081e79245ccb45eaddb9b3cbe2868685628b7304
Related-Bug: #1650089