Add soft delete wrinkle to api-ref
The API reference for deleting a server is making the assumption that
the instance is going to be (hard) deleted from the system, but it
doesn't take into account the case that CONF.reclaim_instance_interval
is set, which if > 0 makes the API 'soft delete' the instance, so that
the instance's vm won't be deleted until the
CONF.reclaim_instance_interval has expired since the instance was soft
deleted. The instance marked as ``SOFT_DELETED`` can be recovered via
``restore`` action before it's really deleted from the system. We should
update the reference to talk about the soft deleted wrinkle so we can set
expectations.

Closes-Bug: #1650328

Change-Id: Ibbb1c225da11234aaac3634769db112a30dfef64