Fix nova-manage cell_v2 discover_hosts RequestContext
In the nova-manage cell_v2 discover_hosts command, the nova.context
module is erroneously being passed to a database API function instead
of a RequestContext object, causing the error on the command line:

  error: 'module' object has no attribute 'session'

This corrects the argument and adds a test assertion for the
RequestContext.

Conflicts:
        nova/tests/unit/test_nova_manage.py

NOTE(mriedem): The conflict is due to the validate transport
tests from ocata not being in test_nova_manage in newton.

Closes-Bug: #1650694

Change-Id: I7b5468218e79ffc77adf15d2f00c063095d62af3
(cherry picked from commit 93a9aaa1cfef8cb2b662790bdae3971368dba808)