conf: Remove 'virt' file
We have "compute" options and "virt" options but there isn't really a
difference between the two. Move two to "network" and the remainder to
"compute", allowing us to drop "virt".

There are no actual changes to the opts, besides a typo fix for the
'reserved_huge_pages' opt.

Change-Id: Ibe7ba5716c6bfb096aabb85c73e5167f337a7327
Implements: blueprint centralize-config-options-ocata