[placement] placement_dev info for testing and gabbi
Add a section to placement_dev.rst explaining how testing of the
placement API works, mostly gabbi. The idea here is not to duplicate
the gabbi docs, but to provide some context on how the use of gabbi
fits in to the existing tests and how to add more, with links off to
the docs for a bit more info.

Change-Id: If99521881a017543acdbbf7e8f1b7f170f5d633d