nova-status: check for compute resource providers
This adds an upgrade check for resource providers which
builds on the cells v2 and placement API checks, so it
changes how upgrade_checks is processed so it's a list
of checks run in order rather than just a dict of things
that can be run independently.

The check added here is relying on the API DB for resource
provider information since Nova does not yet support a
separate placement database.

It also allows for a fresh base install where there are no
resource providers or compute nodes and does not consider
that a failure.

Part of blueprint resource-providers-scheduler-db-filters

Change-Id: I5d316144b3ee63027f3ad5f0984df6cb22f16249