Add connection pool size to vSphere settings
In larger environments, the default value of 10 for the HTTP connection
pool is insufficient as Nova will begin to warn user that the pool
is full and begin discarding connections.

This change adds the ability to configure the connection_pool_size in
Nova.  The default is 10 and the minimum value is also 10.

Closes-Bug: #1494688
Co-Authored-By: rgerganov@vmware.com
Change-Id: I2990899b467ac86664e1b2859f8a190aba7c2905