Fix the display of updated_at time when using memcache driver.
When using mc driver(set servicegroup_driver=mc), the service reports
state by memcache. but we use 'nova service-list' get the status of
services registered in nova, the 'update_at' time is not the last updated
time not as db driver. so we should get the time in memcache to update the
time in response.
Closes-Bug: #1647269

Change-Id: I1d4763436ad6de2483c39f5fe97fa6e8d283d8a3