Make test_quota work with CONF.use_neutron=True by default
There were a couple of nova-network specific floating IP quota
tests in here that needed to be isolated. Plus we need to mock
out the validate_networks network API which for Neutron does a
port quota check when creating instances. Since none of the
tests in here care about that, we just mock it out.

Part of blueprint use-neutron-by-default

Change-Id: Id29326de8337aa0472d537cfe2b8880e90ed7774