Fix assertion in test_instance_fault_get_by_instance
The instance_fault_get_by_instance_uuids DB API method returns
results sorted by created_at in descending order, so the newest
fault should be at the front of the list for a given instance
uuid in the resulting dict.

However, the way the test for this was written the 'expected'
list actually had the faults in order from oldest to newest. The
reason this didn't fail the test was because it was using
_assertEqualListsOfObjects to compare the results, which sorts
the lists before comparing them, which is exactly NOT what we want
when we care about the sort order of the results.

The test is fixed by using _assertEqualOrderedListOfObjects which
leaves the original sort order of the lists to compare and fixes
the 'expected' list by inserting the newest faults at the front
of the list.

Change-Id: I58990194016447b05b42df76141a193595cdcb9c
Related-Bug: #1632247