Make test_server_usage work with CONF.use_neutron=True by default
These tests worked with neutron before d8673cb25 because they
wouldn't load the server security groups extension, but now they
do. So with neutron the server security groups extension tries to
get security groups for the given server(s) being shown even though
that has nothing to do with the point of the server usages API or
this test.

So we add a stub for the security group API just like we have for
the network APIs and use that in this test to ignore any of the
security group output controller handling.

Part of blueprint use-neutron-by-default

Change-Id: Iccf94320a6cbe0fd3b126db310870e93190cc17e