Remove unused init_only kwarg from wsgi app init
Since d8673cb256e17fec05788907fce66898a36ddcf2 the init_only
kwarg is no longer used in the compute API wsgi app setup.
This patch removes it along with it's usage throughout the
compute REST API unit tests where it was used to isolate
testing of extensions, which is not valid anymore as all
extensions are enabled and we removed the ability to
whitelist/blacklist them.

Change-Id: I2f58b7944d0f7d7c05683238474954a65a617ec3