Actually test online flavor migrations
The way the test was written for online flavor migrations
we didn't have any flavors in the main database, so nothing
was migrated and nothing was asserting that we had migrated
anything. This is more or less due to using USES_DB_SELF
and then not populating the main database with any flavors.

This change creates a single flavor in the main database
prior to the online data migration and asserts we actually
started with one and it was migrated.

Change-Id: Ia18f8abdf10785dc396f5c07e239326059528daf
Partial-Bug: #1653261