placement: adds ProviderTree for nested resources
We need a way of representing the hierarchical relationship between
resource providers. This representation needs to be able to work with
either UUIDs or unique string name identifiers due to the way that NUMA
cell and topology objects are constructed in the virt layer.

Following patches show how the ProviderTree interface introduced here
is used.  The scheduler reporting client will build a ProviderTree
containing a resource provider for each hypervisor or baremetal node it
knows about. The information provided to it by the resource tracker
will be used to set inventory information and understand the nesting
relationships between providers. This ProviderTree is used within the
reporting client to set inventory against resource providers contained
in the ProviderTree and determine whether inventory has changed.
Essentially, this ProviderTree object becomes a nova-compute-side cache
of provider and inventory information.

Change-Id: I4e50245a14bb669d409ab489cedea8ce84cf3edc
blueprint: nested-resource-providers