Add 'disabled' to WatchdogAction field
Image property hw_watchdog_action can have a 'disabled'
value which is actually defined in the glance metadefs so
someone using Horizon might pick that. Also, it's the default
behavior in the libvirt driver. However, if you try to create
a server with that set the create fails in nova-api because
the enum field didn't have 'disabled' as a valid value.

This adds it, bumps the ImageMetaProps version and adds tests.

Change-Id: I4cec3e8b8527b909cc60893db26732a19263220d
Closes-Bug: #1633200