Remove the Allocation.create() method.
This method is no longer used, nor should it be. Instead, the
AllocationList.create_all() method should be used to create allocations
in the DB. Tests have been updated to use create_all() instead, and as
create_all() enforces inventory/allocation relationships, several tests
had to be modified so that they properly accounted for this.

Change-Id: I0bccfee155c871babf507cdd1117b5023d5a1361