Cache boot time roles for vendordata.
Some external vendordata services want to provide metadata
based on the role of the user who started the instance. It
would be confusing if the metadata returned changed later
if the role of the user changed, so we cache the boot time
roles and then pass those to the external vendordata
service.

Change-Id: Ieb84c945f4f9a21c2b7b892f9b1ead84dca441e9