Make placement client keep trying to connect
In Newton, placement is optional and computes will stop even trying
to connect to placement when they encounter an error or missing
configuration. We really want them to keep trying so that enabling
the service pre-upgrade does not require restarting all computes
to start filling data.

This patch removes the auto-disable logic and replaces it with a
limited, but persistent warning to the logs about the required
nature of placement for the upgrade. If we had messaged the upcoming
requirement better, I think we could have been less chatty here.
However, we know that it's not been received, so this patch periodically
re-emits the warning and mentions the upgrade specifically.

Conflicts:
        nova/scheduler/client/report.py

NOTE(mriedem): The conflict is due to 3c217acb9 not being in Newton.

Closes-Bug: #1655494
Change-Id: Ie6387afeb239a20d39c00f519e8288f3b3a5e9cb
(cherry picked from commit bbf9b431ee26e4064278e58bf9c177de48e8719a)