Add new configuration option live_migration_scheme
This change adds a new configuration option to change scheme that is used
for live migrations. The change also deprecates old live_migration_uri
in favor of two other flags:

* live_migraton_inbound_addr to configure live migration target URI
* live_migration_scheme to change scheme used for live migration

The new configuration option does not break backward compatibility as
it does not override scheme set in live_migration_uri.

Change-Id: I38b3d1d42b2ffc4a2f6d7398061829d3c745b9a1