Remove redundant arg check in nova-manage cell_v2 verify_instance
argparse will validate that the --uuid is provided, we don't need
an explicit manual check for that in the code, plus the manual
check was violating the --quiet option by printing out an error
message anyway.

This also removes the unit test which we don't really need as
argparse is going to do it's job.

Change-Id: Id1cb79c9be8ff6b94f5a7fcab99e960ad774ceac
Closes-Bug: #1656479