Fix the generated cell0 default database name
This makes the default database name for cell0 be based on the
main database name, not the api one. Before this, we would get
a database name of nova_api_cell0 by default, intead of nova_cell0.
This was confusing because cell0 has "main" schema inside, not "api"
schema.

Conflicts:
        releasenotes/notes/cells-single-migration-command-0e98d66e31e02a50.yaml
        releasenotes/notes/ocata-requires-cellv2-96bd243be874d77f.yaml

NOTE(mriedem): The conflict is due to commit
7c7686366b8c6b887c0d13f202bb66bee32b51a9 not being in Newton.
The change from that commit is left here as it does provide
more useful information about the cell0 mapping and is
relevant to this bug fix. The ocata-requires-cellsv2 release
note from master is deleted in this cherry pick as it's
not relevant to Newton since cells v2 is optional in Newton.

Closes-Bug: #1656673
Depends-On: I4e7f6c5eaa068c98e5c4ef3feaee50d8e4f5d484
Depends-On: Ifdcd6a572dc0b42ba852241e22f6713cd2b084b0
Change-Id: I86797785f76c2c927a4db8fef72b8f8d986af6b9
(cherry picked from commit b28202fb52284dfbc9a9180ac017ddd1977efbf6)